VirtualizedContentTable is now generic
