+++ Src/WitsmlExplorer.Frontend/components/ContentViews/table/tableParts.ts
@@ -15,7 +15,7 @@ export interface ContentTableProps {
   columns: ContentTableColumn[];
   data: any[];
   onSelect?: (row: ContentTableRow) => void;
-  onContextMenu?: (event: React.MouseEvent<HTMLLIElement | HTMLTableRowElement, MouseEvent>, selectedItem: Record<string, any>, checkedItems: Record<string, any>[]) => void;
+  onContextMenu?: (event: React.MouseEvent<HTMLElement, MouseEvent>, selectedItem: Record<string, any>, checkedItems: Record<string, any>[]) => void;
   checkableRows?: boolean;
 }
